Move boot module lookup into registry
