+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/hub/registry/BootClassRegistry.java
@@ -34,7 +34,9 @@ import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.ProviderNotFoundException;
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -77,6 +79,10 @@ public final class BootClassRegistry extends AbstractRuntimeClassRegistry {
     /// `-Xbootclasspath/a:` entry (e.g. `/path/to/boot-append.jar`) that supplied the first
     /// successfully defined class in the package.
     private static final ConcurrentHashMap<String, String> loadedBootAppendPackageLocations = new ConcurrentHashMap<>();
+
+    /// Lazily initialized map of a package name to the name of the
+    /// boot-loader loaded module whose descriptor [includes the package][java.lang.module.ModuleDescriptor#packages()].
+    private static volatile Map<String, String> bootPackageNameToModuleName;
 
     @Platforms(Platform.HOSTED_ONLY.class)
     public BootClassRegistry() {
@@ -162,8 +168,30 @@ public final class BootClassRegistry extends AbstractRuntimeClassRegistry {
     }
 
     /// Finds the boot module that declares `internalPackageName`.
+    public static String bootModuleNameForPackage(String internalPackageName) {
+        String packageName = internalPackageName.replace('/', '.');
+        if (bootPackageNameToModuleName == null) {
+            synchronized (BootClassRegistry.class) {
+                if (bootPackageNameToModuleName == null) {
+                    Map<String, String> result = new HashMap<>();
+                    for (Module m : ModuleLayer.boot().modules()) {
+                        if (m.getClassLoader() == null) {
+                            for (String p : m.getDescriptor().packages()) {
+                                result.put(p, m.getName());
+                            }
+                        }
+                    }
+                    // Create an immutable copy of the map
+                    bootPackageNameToModuleName = Map.copyOf(result);
+                }
+            }
+        }
+        return bootPackageNameToModuleName.get(packageName);
+    }
+
+    /// Finds the boot module that declares `internalPackageName`.
     static String getBootModuleForPackage(String internalPackageName) {
-        return BootLoaderPackageAccess.bootModuleNameForPackage(internalPackageName);
+        return bootModuleNameForPackage(internalPackageName);
     }
 
     /// Finds the boot module for `internalPackageName` after the package is defined to the boot
